# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_oauth_interfaces_api_handlers.OAuthConnectionResponse:
    properties:
      connected_at:
        type: string
      id:
        type: string
      provider:
        type: string
      provider_host:
        type: string
      provider_username:
        type: string
      scopes:
        items:
          type: string
        type: array
    type: object
...
  opscore_backend_internal_oauth_domain.OAuthCallbackRequest:
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      code:
        type: string
      gitlabUrl:
        type: string
      provider:
        $ref: '#/definitions/opscore_backend_internal_oauth_domain.Provider'
      state:
        type: string
    required:
    - code
    - provider
    - state
    type: object
  opscore_backend_internal_oauth_domain.OAuthTokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
  opscore_backend_internal_oauth_domain.Provider:
    enum:
    - github
    - gitlab
    - gitlab-self-hosted
    type: string
    x-enum-varnames:
    - ProviderGitHub
    - ProviderGitLab
    - ProviderGitLabSelfHosted
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the OpsCore backend service.
  title: OpsCore Backend API
  version: "1.0"
paths:
...
  /auth/oauth/callback:
    post:
      consumes:
      - application/json
      description: Exchanges OAuth authorization code for access token
      parameters:
      - description: OAuth callback request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_oauth_domain.OAuthCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/opscore_backend_internal_oauth_domain.OAuthTokenResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: OAuth callback endpoint
      tags:
      - OAuth
  /auth/oauth/connect:
    post:
      consumes:
      - application/json
      description: Exchanges OAuth authorization code for access token and saves to
        database
      parameters:
      - description: OAuth callback request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_oauth_domain.OAuthCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_oauth_interfaces_api_handlers.OAuthConnectionResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: OAuth callback endpoint with token storage
      tags:
      - OAuth
  /auth/oauth/connections:
    get:
      description: List all OAuth connections for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List OAuth connections
      tags:
      - OAuth
  /auth/oauth/connections/{connectionId}/repositories:
    get:
      description: List all repositories accessible via a specific OAuth connection
      parameters:
      - description: OAuth connection ID
        in: path
        name: connectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid connection ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List repositories by OAuth connection
      tags:
      - OAuth
  /auth/oauth/connections/{connectionId}/repositories/{owner}/{repo}/contents:
    get:
      description: Get files and directories at a specific path via OAuth connection
      parameters:
      - description: OAuth connection ID
        in: path
        name: connectionId
        required: true
        type: string
      - description: Repository owner
        in: path
        name: owner
        required: true
        type: string
      - description: Repository name
        in: path
        name: repo
        required: true
        type: string
      - description: Directory path (empty for root)
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get repository contents
      tags:
      - OAuth
  /auth/oauth/connections/{connectionId}/repositories/{owner}/{repo}/files/{filePath}:
    get:
      description: Get the content of a specific file via OAuth connection
      parameters:
      - description: OAuth connection ID
        in: path
        name: connectionId
        required: true
        type: string
      - description: Repository owner
        in: path
        name: owner
        required: true
        type: string
      - description: Repository name
        in: path
        name: repo
        required: true
        type: string
      - description: File path
        in: path
        name: filePath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get file content
      tags:
      - OAuth
  /auth/oauth/connections/{provider}:
    delete:
      description: Remove an OAuth connection for the authenticated user
      parameters:
      - description: Provider name (github, gitlab)
        in: path
        name: provider
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid provider
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Disconnect OAuth provider
      tags:
      - OAuth
...
securityDefinitions:
  BearerAuth:
    description: 'Enter "Bearer " followed by your JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"